# By Dominik Kapusta (41) and others
# Via Dominik Kapusta (9) and others
* main: (138 commits)
  Make sure when we set custom config url, we don't expect etag in return (#1994)
  Add PixelKit source parameter (#1989)
  Fix internal user toggling (#2000)
  Show alert and display warning icon in Sync Settings when data syncing is disabled (#1996)
  DBP: Integrate subscription account authentication to DBP (#1995)
  Improve bookmarks html reader (#1986)
  Add Sync feature flags (#1992)
  Add daily stats pixel (#1993)
  Do not reload DBP tab when switching to it (#1942)
  Fix: external application requests via redirect URLs shows wrong origin. (#1900)
  Update clean-app.sh to work on macOS Sonoma and include NetP containers (#1988)
  Fix: "SwiftLintPlugin" must be enabled before it can be used (#1987)
  Prevent VPN server list persistence failures (#1985)
  add test can remove data (#1980)
  Remove VPN upgrade card (#1983)
  Fix low-res VPN warning asset (#1984)
  DBP: Fix unreliable date tests (#1981)
  Add search retention pixel for NetP (#1964)
  Sabrina/sync e2e tests (#1959)
  swiftlint build plugin (#1318)
  ...

# Conflicts:
#	DuckDuckGo.xcodeproj/project.pbxproj
#	DuckDuckGo.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
#	DuckDuckGo/Application/AppDelegate.swift

+++ DuckDuckGo/NetworkProtection/AppTargets/BothAppTargets/NetworkProtectionDebugUtilities.swift
@@ -38,10 +38,15 @@ final class NetworkProtectionDebugUtilities {
 
     private let loginItemsManager: LoginItemsManager
 
+    // MARK: - Settings
+
+    private let settings: VPNSettings
+
     // MARK: - Initializers
 
-    init(loginItemsManager: LoginItemsManager = .init()) {
+    init(loginItemsManager: LoginItemsManager = .init(), settings: VPNSettings = .init(defaults: .netP)) {
         self.loginItemsManager = loginItemsManager
+        self.settings = settings
 
         let ipcClient = TunnelControllerIPCClient(machServiceName: Bundle.main.vpnMenuAgentBundleId)
 
@@ -51,8 +56,14 @@ final class NetworkProtectionDebugUtilities {
 
     // MARK: - Debug commands for the extension
 
-    func resetAllState(keepAuthToken: Bool) async throws {
-        networkProtectionFeatureDisabler.disable(keepAuthToken: keepAuthToken, uninstallSystemExtension: true)
+    func resetAllState(keepAuthToken: Bool) async {
+        let uninstalledSuccessfully = await networkProtectionFeatureDisabler.disable(keepAuthToken: keepAuthToken, uninstallSystemExtension: true)
+
+        guard uninstalledSuccessfully else {
+            return
+        }
+
+        settings.resetToDefaults()
 
         NetworkProtectionWaitlist().waitlistStorage.deleteWaitlistState()
         DefaultWaitlistActivationDateStore().removeDates()
@@ -63,16 +74,16 @@ final class NetworkProtectionDebugUtilities {
     }
 
     func removeSystemExtensionAndAgents() async throws {
-        await networkProtectionFeatureDisabler.removeSystemExtension()
+        try await networkProtectionFeatureDisabler.removeSystemExtension()
         networkProtectionFeatureDisabler.disableLoginItems()
     }
 
     func sendTestNotificationRequest() async throws {
-        await ipcClient.debugCommand(.sendTestNotification)
+        try await ipcClient.debugCommand(.sendTestNotification)
     }
 
-    func expireRegistrationKeyNow() async {
-        await ipcClient.debugCommand(.expireRegistrationKey)
+    func expireRegistrationKeyNow() async throws {
+        try await ipcClient.debugCommand(.expireRegistrationKey)
     }
 }
